feat: Don't require backwards verification to add a contact to verified group
All UIs should check this anyway and explain to the user why an only forward-verified contact mayn't
be added to verified groups (to protect from situations when the contact doesn't trust our key and
rejects the addition). In Core it looks superfluous (in fact we just don't know what happens, maybe
the contact has gone offline, but otherwise everything is fine), so let's only check forward
verification, anyway the check is raceful and at the point of the function return the verification
might break.