fix: Prevent accidental wrong-password-notifications
Over the past years, it happend two times that a user came to me worried about the "Cannot
login as ***. Please check if the e-mail address and the password are
correct." message.

I'm not sure why this happened, but I think we should make the logic for
showing this notification stricter:
- Before: The notification is shown if connection fails two times in a row, and
  the second error contains the word "authentication".
- Now: The notification is shown if the connection fails two times in a row,
  and _both_ error messages contain the word "authentication".

The second commit just renames `login_failed_once` to
`authentication_failed_once` in order to reflect this change.