fix: skip IDLE if we got unsolicited FETCH
This may indicate that there was a new \Seen flag
that we don't want to skip.

Also don't drain unsolicited responses while scanning folders.
Now we only drain unsolicited responses right before IDLE
and always redo the whole fetch cycle if there have been some.
Some message in the scanned folder may not be fetched
that would be previously fetched otherwise,
but it will be picked up on the next folder scan.