never notify SELF
it may be handy for an xdc to have only one list of all adresses,
or there may just be bugs.

in any case, do not notify SELF, e.g. in a multi-device setup;
we're also not doing this for other messages.

this is also a preparation for having an option to notify ALL.