test: Fix panic in `receive_emails` benchmark
The benchmark function (e.g. `recv_all_emails()`) is executed multiple times on the same context.
During the second iteration, all the emails were already in the database, so, receiving them again failed.

This PR fixes that by passing in a second `iteration` counter that is
different for every invocation of the benchmark function.