api: cffi: add `dc_event_get_json` which gives the jsonrpc representation of an event
This has the advantage that fields are named/labled and there can be potentially more than 2 (data1 & data2) fields on the events.

This removes the need for the (potentialy confusing) overloading data1 with `dc_event_get_data1_str` to add more information to events.
And also allows gradual/partial moving to the jsonrpc api, when needed or wanted.