fix: Emit MsgsNoticed on receipt of an IMAP-seen message
`imap::Session::sync_seen_flags()` emits `MsgsNoticed` for existing messages seen on other devices,
so `receive_imf` should do the same when it receives a seen message. Otherwise a multi-device user
may see a new message notification on device A, just swipe it, then see another new message
notification and mark it as read, and when their device B goes online, it will show a notification
for the first message, and it won't be removed because `MsgsNoticed` isn't emitted. I have checked
this with my DC Android and Desktop. With this fix the notification can be removed if UIs handle
`MsgsNoticed` accordingly, though currently i can't achieve this behaviour on Desktop. Anyway, the
fix unifies the `receive_imf` behaviour with `sync_seen_flags()`.