fix: Don't emit MsgsNoticed when a message was seen on IMAP
`receive_imf` doesn't emit `MsgsNoticed` on receipt of an IMAP-seen message, so the code was
inconsistent. The easiest fix is to not emit `MsgsNoticed` when a message was seen on another
device, we don't want to emit `MsgsNoticed` if not all messages are seen, otherwise that removes
useful notifications from UIs and may lead to missing messages. Checking that all previous chat
messages are at least `InNoticed` looks complicated, let's not do this for now.