remove get_original_msg_id()
agree with @Hoccuri after thinking it over,
the limited cornercase of deleted message but not deleted chat
seems not to be worth additional effort and states -
and may even worsen UX as it is unclear what happens when tapping the
"goto original" button (or one need to use different icons or sth)

if we want to do sth. about that,
it might be better to go the way we went for quotes:
so, eg. keep a copy of the chatname.
but let's see if this is really needed and an issue at scale.